requirements.txt
numpy==1.19.5
matplotlib==3.3.3
tensorflow==2.4.0
deepface==0.0.49
opencv_python==4.5.1.48
face_recognition==1.3.0
pyscreenshot==2.3
Pillow==8.1.0
git status